Move wallet tasks to test_wallets
