+++ tests/test_wallets.py
@@ -1,7 +1,12 @@
 # pragma pylint: disable=missing-docstring
+from copy import deepcopy
 from unittest.mock import MagicMock
 
-from tests.conftest import get_patched_freqtradebot
+import pytest
+
+from freqtrade.constants import UNLIMITED_STAKE_AMOUNT
+from freqtrade.exceptions import DependencyException
+from tests.conftest import get_patched_freqtradebot, patch_wallet
 
 
 def test_sync_wallet_at_boot(mocker, default_conf):
@@ -106,3 +111,55 @@ def test_sync_wallet_missing_data(mocker, default_conf):
     assert freqtrade.wallets._wallets['GAS'].used is None
     assert freqtrade.wallets._wallets['GAS'].total == 0.260739
     assert freqtrade.wallets.get_free('GAS') == 0.260739
+
+
+def test_get_trade_stake_amount_no_stake_amount(default_conf, mocker) -> None:
+    patch_wallet(mocker, free=default_conf['stake_amount'] * 0.5)
+    freqtrade = get_patched_freqtradebot(mocker, default_conf)
+
+    with pytest.raises(DependencyException, match=r'.*stake amount.*'):
+        freqtrade.wallets.get_trade_stake_amount('ETH/BTC', freqtrade.get_free_open_trades())
+
+
+@pytest.mark.parametrize("balance_ratio,result1", [
+                        (1, 0.005),
+                        (0.99, 0.00495),
+                        (0.50, 0.0025),
+])
+def test_get_trade_stake_amount_unlimited_amount(default_conf, ticker, balance_ratio, result1,
+                                                 limit_buy_order_open, fee, mocker) -> None:
+    mocker.patch.multiple(
+        'freqtrade.exchange.Exchange',
+        fetch_ticker=ticker,
+        buy=MagicMock(return_value=limit_buy_order_open),
+        get_fee=fee
+    )
+
+    conf = deepcopy(default_conf)
+    conf['stake_amount'] = UNLIMITED_STAKE_AMOUNT
+    conf['dry_run_wallet'] = 0.01
+    conf['max_open_trades'] = 2
+    conf['tradable_balance_ratio'] = balance_ratio
+
+    freqtrade = get_patched_freqtradebot(mocker, conf)
+
+    # no open trades, order amount should be 'balance / max_open_trades'
+    result = freqtrade.wallets.get_trade_stake_amount('ETH/BTC', freqtrade.get_free_open_trades())
+    assert result == result1
+
+    # create one trade, order amount should be 'balance / (max_open_trades - num_open_trades)'
+    freqtrade.execute_buy('ETH/BTC', result)
+
+    result = freqtrade.wallets.get_trade_stake_amount('LTC/BTC', freqtrade.get_free_open_trades())
+    assert result == result1
+
+    # create 2 trades, order amount should be None
+    freqtrade.execute_buy('LTC/BTC', result)
+
+    result = freqtrade.wallets.get_trade_stake_amount('XRP/BTC', freqtrade.get_free_open_trades())
+    assert result == 0
+
+    # set max_open_trades = None, so do not trade
+    freqtrade.config['max_open_trades'] = 0
+    result = freqtrade.wallets.get_trade_stake_amount('NEO/BTC', freqtrade.get_free_open_trades())
+    assert result == 0
